Generator for LoadContinentEvent
Process function process the event from p2app.views, then sources to ContinentEvents class in function_import module which returns the values of the contents in the selected continent. That Continent tuple is then yielded with a ContinentLoadedEvent tag back to p2app.views. Tested with tkinter GUI for correctness.