Generator Filtering for SaveContinentEvent
Generator filters for saving edited continent and routes to the ContinentEvents class that runs a try-except to save continent. The class returns a continent tuple of the new ones or returns an error message why it can't save. Results are yielded back to views module. Tested with tkinter GUI using various values.