Generator Filtering for Loading Country
Generator filters for specific event name and then routes to CountryEvent class in engine/function_imports module. The class filters for the country_code and then returns a Country tuple of the specific country. The Country tuple is then yielded in the engine/main module to the views modules. Code tested manually with Tkinter GUI.