Generator Filtering for Creating New Country
Generator filters for specific event name and then routes to CountryEvent class in engine/function_imports module. The class tries to insert country into the table, if failure a exception message is returned, otherwise a country tuple is returned. Depending on outcome, the content of either tuple or reason is yielded to views modules. Manual Tests run in tkinter GUI.